requirements.txt
google_auth_oauthlib
pyreadline
urllib3
setuptools
httplib2shim
progress
google-api-python-client
python-telegram-bot
